kernelize/backend/requirements.txt
# KERNELIZE Knowledge Compression Platform - Production Requirements
# ================================================================

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database & ORM
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.12.0

# Authentication & Security
python-jose[cryptography]>=3.3.3
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# AI & Machine Learning
torch>=2.1.0
transformers>=4.35.0
sentence-transformers>=2.2.2
numpy>=1.26.0
scikit-learn>=1.3.0

# Multi-Modal Processing
Pillow>=10.0.0
opencv-python>=4.8.0
librosa>=0.10.1
whisper>=1.1.10

# Data Processing
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dateutil>=2.8.2

# Monitoring & Metrics
prometheus-client>=0.19.0
psutil>=5.9.0

# Configuration & Logging
pyyaml>=6.0.1
python-dotenv>=1.0.0
loguru>=0.7.2

# HTTP Client
httpx>=0.25.0
aiofiles>=23.2.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Code Quality
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Development Tools
ipython>=8.17.0
jupyter>=1.0.0
